requirements.txt
flask
requests
python-dotenv
huggingface_hub
Pillow